lesson21_web/requirements.txt
flask
gunicorn
google-genai
line-bot-sdk